requirements.txt
streamlit
web3
py-solc-x
qrcode
pillow
pyzbar
opencv-python
